Update Libraries and Application
Commit updates all libraries and adjusts the application to be able to work with them. Important to notice is the shift from Vue 2 to Vue 3. Deprecated functions and components have been replaced with newer versions, replacements or custom variants. Tested locally with: npm run serve, on port 8080.

Commits can be added to fix minor issues that were overseen, like: colors, library issues or functional traits that do not work.